feat: Implement backend structure for ML, simulation, and AI tutoring
- Added Flask backend with routes for ML and simulation functionalities.
- Created endpoints for training and testing regression and classification models.
- Integrated simulation capabilities with predefined physics equations.
- Developed a recommendation system for model selection based on dataset properties.
- Enhanced frontend components for simulation and home pages.
- Established a structured plan for future upgrades and features.
