+++ backend/routes/ai_routes.py
@@ -1,0 +1,6 @@
+# AI Routes
+from flask import Blueprint
+
+ai_bp = Blueprint('ai', __name__)
+
+# Import controllers here or define routes
