+++ frontend/src/Hero.jsx
