+++ frontend/src/Pages/SimulationPage.jsx
