+++ frontend/src/componants/MathAI.jsx
