+++ frontend/src/componants/Simulation.jsx
